Move LogDataBOUTdecoder to decoder.py
